Unfinished commit

+++ 2501ButtonsTestHW/RITViewController.h
@@ -16,7 +16,7 @@
 @property (weak, nonatomic) IBOutlet UILabel *displayLabel;
 @property (strong, nonatomic) IBOutletCollection(RITCalcButton) NSArray *operationButtons;
 
-@property (assign, nonatomic) long long firstValue;
-@property (assign, nonatomic) long long secondValue;
+@property (assign, nonatomic) double firstValue;
+@property (assign, nonatomic) double secondValue;
 
 @end
